Tests: Add verbose info if `submit_and_wait` times out
If the `submit_and_wait` fixture times out waiting for the submitted
process to reach the desired state, usually there is a problem with the
daemon workers. To make debugging easier, the status of the daemon as
well as the content of the daemon log file are included in the exception
message.