Engine: Do not call serializer for `None` values (#5694)
Any port that defines a `serializer`, will have that serializer invoked
when a value is passed to this port. The invokation was correctly
skipped if no serializer was defined or if the value already was a
`Data` node (in which case serialization is no longer needed), but it
was also called when this value is `None`.

It arguably doesn't make sense to serialize a value that has not been
defined. This behavior would cause problems for processes that define a
port with a serializer but that also accept `None`. In this case the
serializer would nevertheless be called and, in the case of the
`to_aiida_type` serializer for example, would except.