Dependencies: Update requirements for `kiwipy` and `plumpy`
The new version `kiwipy==0.8` and `plumpy==0.22` provide compatiblity
with newer versions of `aio-pika==8.0` which comes with various
connection stability improvements.

The only known problem is that `Communicator.close()` times out if at
least one process has been run. A test is added to capture this behavior
in `tests/manage/test_manager.py` which currently fails as a
`TimeoutError` is thrown. A lot of debugging has not yet led to finding
the cause nor a solution.

Since this behavior only seems to be appearing in the tests and does not
seem to affect regular usage, the upgrade is accepted regardless.