Tests: Suppress logging and warnings from temporary profile fixture
When the test suite is run without specifying a particular test profile,
one is created on the fly using the `TemporaryProfileManager`. This
would create a temporary configuration directory and a temporary profile
which would generate a `UserWarning` and some `REPORT` level log
messages.

This info causes unnecessary noise and the warning even causes `pytest`
to record that the tests issued a warning, but the warning is completely
to be expected and benign, so here we capture the warnings and log
messages.