DevOps: Remove `--use-feature` flag in `pip install` of CI (#5703)
The `tests` job of the CI workflow used the `--use-feature=2020-resolver`
flag in the `pip install` call to force using a specific package
dependency resolver. We adopted this way back when `pip` introduced this
new dependency resolver to be early-adopters to detect problems as soon
as possible. In latest versions of `pip` this option is no longer
available since the resolver is now the default, so we simply remove the
flag.