Add the `ContainerizedCode` data plugin
This implementation of the `AbstractCode` interface allows running a
`CalcJob` within a container on a target computer. The data plugin
stores the name of the container image, the executable within that
container, the command with which the container should be launched and
the `Computer` on which the container can be run.