`Code`: Forward `get_execname` method to `get_executable`
The `get_execname` method was renamed to `get_executable` on the
`AbstractCode` base class in the recent refactor of the `Code` class.
However, certain plugin packages used this method. Since the new
concrete classes `InstalledCode` and `PortableCode` inherit from the
legacy `Code` class, they do have the method, but not the associated
attributes and so will return an empty string. The fix is to have the
`Code.get_execname` class call the `get_executable` method and cast the
`pathlib.Path` value that is returned to a `str` to match the old
interface.