DevOps: Change `on` trigger for `ci-code` and `test-install` workflows
The reason is that Github recently stopped making the `GITHUB_TOKEN`
available in workflows triggered by pull requests from forks for
security reasons. As a result, the `peter-evans/commit-comment@v2`
action which is used to add a commit to the PR in case a particular
check failed, fails to execute since it doesn't have permissions.
Changing the workflow trigger to `pull_request_target` works around this
problem as the same limitation does not hold there. For more info:

https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#pull_request_target