`Process`: Add hook to customize the `process_label` attribute
Up till now the `process_label` of a `ProcessNode` instance representing
the execution of a `Process` would be set to the name of the process
class. This was hardcoded and could not be changed, but in certain cases
it may be useful for the subclass to provide a more meaningful label.

Here the `_build_process_label` method is called which should return the
process label. It can be overridden by subclasses to return another
string. When an instance of the `Process` is run, the value returned by
this method will be stored in the `process_label` attribute.