Mark relevant `Process` exit codes as `invalidates_cache=True`
The `Process` and `CalcJob` classes define various exit codes, all of
which will correspond to processes that should not be used as a source
for caching. Therefore, these exit codes should properly declare
`invalidate_cache=True` which will automatically prevent any node with
that exit code from being cached from.

The `ArithmeticAddCalculation` and `TemplatereplacerCalculation plugins
are also updated to properly mark exit codes that should not be cached
from.