Add a client to connect to RabbitMQ Manamegement HTTP API
RabbitMQ provides a plugin `rabbitmq_management` that provides an HTTP
API to manage a RabbitMQ instance. For example, it can be used to
inspect existing queues, create new ones, or delete queues.

For the client to work, the plugin should first be enabled with:

    sudo rabbitmq-plugins enable rabbitmq_management

If the API cannot be connected to a `ManagementApiConnectionError`
exception is raised.

In the Github Action workflows, the command above won't work, since
RabbitMQ is installed as a service in a Docker container. However, it
can be enabled in the Docker container by adding the `-management`
suffix to the version specification of the RabbitMQ service. Since the
API is exposed on port 15672, that is also added to the forward list.