RabbitMQ: Remove support for v3.5 and older
RabbitMQ 3.5 has been EOL since 31 October 2016:

    https://www.rabbitmq.com/versions.html

This version required to explicitly run `support_deprecated_rabbitmq` on
the `pamqp` library (which is used all the way down the stack by
`aiormq` to handle interactions with RabbitMQ) to enable compatibility.
In removing this, the explicit dependency on `pamqp` is also removed.

The `aiida.manage.manager.is_rabbitmq_version_supported` function which
is used to check compatibility when a profile is loaded is updated to
include the lower bound.