CLI: Add the `verdi devel rabbitmq` command group
The command group has three subcommands:

 * `server-properties`: Report properties of RabbitMQ server
 * `queues`: Interact with RabbitMQ queues
 * `tasks`: Deal with messages stored in tasks queues

The first command is a simple utility that can come in handy during
debugging as it reports various server settings, more so than the
version of RabbitMQ that is reported by `verdi status`.

The `queues` command is also mostly useful for debugging. It currently
allows to list, create and delete queues.

The most important command is the `tasks` command. Currently, there is a
bug where it is possible that tasks in the process queue (which are
consumed by daemon workers to continue submitted processes) can get
lost, despite the persistence requirements on the queue and the
messages. When a tasks gets lost, the corresponding process will become
a "zombie" and seems stuck. Currently there was no easy way to confirm
that a process had become a zombie due to its task having been lost.

The `list` method fixes this problem as it will introspect all task
messages in the process queue and collect the process ids that they
correspond to. This can be used to see which active processes no longer
have a corresponding task and so are a zombie.

The `analysis` command is a utility command that automates this process.
It will get all process ids of the existing tasks in the process queue
and cross-reference those with the active processes as stored in the
database of AiiDA's storage backend. By default it reports if there are
inconsistencies between the two sets. If there are, the user can run the
command again with the `--fix` flag and the command will automatically
discard tasks that are incorrect and recreate those that are missing.