Tests: Use `pytest-regression` for sphinx extension test
The tests for the Sphinx extension for process specifications worked
using a direct comparison of the parsed XML file. Although this worked
fine, we have the `pytest-regressions` plugin for this which handles
this automatically, including a useful switch `--force-regen` to
regenerate the reference file if it is supposed to change.