`PsqlDosBackend`: ensure sqla sessions are garbage-collected on `close`
When calling `Manager.unload_profile`, the `close` method is called on
the `StorageBackend`. For the `PsqlDosBackend` this means cleaning up
all SqlAlchemy connections and sessions. This was implemented, however,
SqlAlchemy would often still be holding on to at least one session in
the `WeakrefValueDictionary` of `sqlalchemy.orm.sessions._sessions`. It
turns out however that the reference was already cleared in fact, as
calling the garbage collector explicitly cleans the session references.