fix: store user info early
`verdi setup` and `verdi quicksetup` store information about the user,
such as their name, email, and institution, but they only do so when the
setup finishes successfully.
This is tedious when setup fails as the user is forced to re-enter
this information anew for every try. Here we simply make sure to store
this information early.