ORM: Return `None` in `get_function_source_code` instead of excepting
The `FunctionCalculationMixin` which is used for the nodes of process
functions has the function `get_function_source_code` which will raise
`FileNotFoundError` if the source code file does not exist in the
repository. This will happen typically for functions defined in an
interactive shell where the `inspect` package cannot determine the
source code.

The docstring erroneously said that `None` would be returned, but even
worse it said that the absolute filepath would be returned, whereas it
was supposed to give the source code as a string.

The docstring is fixed and the method now returns `None` if the source
code file doesn't exist instead of raising.