`QueryBuilder`: Fix bug when mutating nodes during `iterall`
While iterating over a collection of nodes returned by the
`QueryBuilder.iterall` method, trying to mutate the state of a node, for
example by storing an extra, would raise an exception. The reason is
that the `ModelWrapper` calls `commit` on the session once a field of a
stored node is mutated, unless the session is within a transaction. In
most user use-cases, however, where they are iterating over nodes to set
some extras, they do no open an explicit transaction and we cannot
expect them to, and so the code excepts.

This bug has existed for the sqlalchemy backend going as far back as at
least v1.6.9 (which is confirmed) but probably it has been there for
much longer. For Django this used to work, but that was dropped in v2.0.

The reason for the exception is that the cursor that is being used for
iterating over the rows of the query is no longer valid. This is an
error raised by the DBAPI, which is psycopg2 in our case, and occurs
often when commiting on a cursor that is in the middle of fetching rows
from a query and this is therefore strongly discouraged.

One solution is to remove the explicit commit in the `ModelWrapper.save`
method. To still ensure that changes to stored nodes get flushed, we
enable `autoflush=True` on the `sessionmaker`. This is a recommended way
of letting the engine control that changes are flushed to the database
whenever it needs to.

However, this change will have consequences for other usage though that
will render it unusable. For example, it would become possible for a
user to load a node in an interactive shell, change its state and exit
immediately. The autoflush will not have had the opportunity to flush
the changes to the database and the changes would be lost.

The solution chosen here is to open a transaction in the `QueryBuilder`
endpoints that fetch rows using `yield_per`, i.e., in `iterall` and
`iterdict`. This nested session will prevent `ModelWrapper.save` from
calling `commit` on the cursor when a node is mutated during the
iteration over the row collection.