CLI: Add the `devel launch-add` command (#5733)
This command will launch an `ArithmeticAddCalculation`. By default the
job is run blockingly, but with the `--daemon` switch it can be sent to
the daemon instead.

This command is useful for debugging purposes as it requires zero setup.
All that is needed is an AiiDA profile. The command will automatically
configure the localhost computer, if not already done, and create a
`Code` instance for `bash`, likewise if it doesn't already exist. This
makes it very useful when helping users to debug their installation as
with a single command and no setup it can be verified whether there is
a problem with the installation or with the plugins they are running.
Also for development it is easy in scenarios where arbitrary calculation
jobs need to be run.