`QueryBuilder`: Add test that excepts when mutating nodes in `iterall`
Currently, when iterating over the rows of a query using `iterall` or
`iterdict`, will raise `sqlalchemy.ProgrammingError` if any of the nodes
returned by the iterator are mutated, for example by setting an extra.

The reason is that adding an extra is mutating the node, and the model
is wrapped in the `ModelWrapper` class which will automatically call
save if a field of a stored node is changed. This is calling `commit` on
the session which is invalidating the cursor. The documentation of
sqlalchemy explicitly states that one should not commit on a cursor that
is being used to iterate over a query with `yield_per`.