🐛 FIX: Import archive into large DB
As detailed in https://www.sqlite.org/limits.html, SQLITE_MAX_VARIABLE_NUMBER puts a limit on how many variables can be used in a single SQL query.
This size can be easily reached, if filtering by nodes in a large database.
Therefore, this commit changes the filtering of UUIDs to be on the client side,
then batches queries for the full nodes by a fixed number.