Process functions: Add the `get_source_code_function` method (#4554)
This method returns the source code of just the process function itself,
i.e., the wrapped function plus the decorator itself.

The existing method `get_function_source_code` would return the source
code of the entire file in which the process function was defined, but
this can at times include a lot of other code that is not always useful
to have. The `get_function_source_code` method is deprecated and replaced
by the `get_source_code_file` method to have consistent naming.

The `get_source_code_function` implementation retrieves the source code
of the function by calling `get_source_code_file` and extracting just
the lines of the function. This is accomplished by using the starting
line of the function, which already used to be stored, combined with the
total number of lines that make up the function, which is an attribute
that is stored from now on for process functions.