Process functions: Add serialization for Python base type defaults
As of 06802099fad32fb113b52fca40c867674aa7ea94, process functions will
automatically add the `to_aiida_type` serializer for arguments such that
Python base types are automatically converted to the corresponding AiiDA
data type if not already a `Data` instance.

The same was not added for defaults however, so if a user defined a
calcfunction with a Python base type default, an exception would be
raised once the dynamic `FunctionProcess` would be constructed. The
default would be passed to the `InputPort` constructor as is and so
would not validate against the type check.

Here we update the dynamic process spec generation to detect if a
default is provided for an argument and if so serialize it to the node
equivalent. Note that this is done indirectly through a lambda as using
constructed node instances as defaults in process specs can cause issues
for example when the spec is exposed into another process spec, the
ports are deep copied, including the defaults of the ports and deep
copying of data nodes is not supported.