CLI: Improve the quality of `verdi code list` output
The output of `verdi code list` was being formatted "by hand" and as a
result was quite different from the output of other commands that
typically use the `tabulate` library to do the formatting.

The command is refactored to use `tabulate`. A single query is created
which is more readable compared to the old logic and filters for each of
the entities are built up consistently based on the options passed to
the command.

The `-P/--project` option is added. Just as for `verdi proces list` this
option allows to control which properties of the query are projected and
printed to stdout. This means that the `-o/--show-owner` option is no
longer needed and so is deprecated. It can be replaced by using `-P user`
with the new option.

The new option also allows projecting the "entry_point" of the `Code`
instances that are matched by the query. This has become interesting
since recently the `AbstractCode` class allows plugins to provide
custom implementations registered through an entry point.

In adding the project option, the `-P/--input-plugin` option had to be
changed since the short flag `-P` overlapped. This was anyway a good
idea since the attribute `input_plugin` has long ago been renamed to
`default_calc_job_plugin` so the option now takes the flags `-d` and
`--default-calc-job-plugin`.