Fixtures: Make `daemon_client` session scoped
Shutting down the daemon after every test is costly, especially if many
tests require a daemon and so each test will start and stop the daemon
as opposed to it being shut down once at the end of the session.

The `daemon_client` was initially introduced as being session scoped but
then changed to being function scoped because there was a subtle bug in
that caused the `test_duplicate_subscriber_identifier` to fail. It only
failed though when `test_revive` of `engine/processes/test_control.py`
and `tests/engine/processes/calcjobs/test_calc_job.py` were run before
it.

The problem is that these tests use the `aiida_profile_clean` fixture
which reset the storage backend. However, it doesn't restart the daemon
and it may still hold on to references to entities in the database that
will have been cleaned. The solution is to also restart the daemon if it
was running when resetting the profile. This bug was solved in the
previous commit and so now the original ideal intended scope of the
`daemon_client` fixture can be reinstated.