`ProfileManager`: restart daemon in `clear_profile`
When calling `clear_profile` the storage backend will be emptied. If a
daemon is running that is connected to that storage, it may still be
holding references to entities that will now have been removed. As soon
as the daemon will then pick up another process in a later test, it will
hit an exception trying to access the entity in the database that no
longer exists.

To solve this, the `clear_profile` now checks whether the daemon is
running before clearing the storage in which case it stops it. The
daemon is not restarted because this is a costly operation. If there are
many tests that clear the profile but that do not need the daemon, it is
restarted for each test unnecessarily, greatly increasing test run time.
Any test that requires the daemon to be started, should make sure of
that itself.