Tests: Replace all use of `aiida_profile_clean` with `aiida_profile`
The `aiida_profile_clean` will clean the storage backend, which often is
an expensive operation. For the `PsqlDosBackend` for example, this means
dropping all database tables and deleting the file repository before
recreating a clean instance of both from scratch.

Often tests just require a profile to be loaded but it doesn't have to
be clean. The tests can be sped up significantly by only using the
`aiida_profile_clean` fixture where it is really necessary.