`BaseRestartWorkChain`: Fix bug in `_wrap_bare_dict_inputs` (#5757)
In ae79c89ecb0f8c72b6429c4a453408d02dffc471, released with `v2.1.0`, the
`InputPort` class was updated to automatically add `None` as a valid
type to the `valid_type` tuple if the port is not required. This
unintentionally broke the `BaseRestartWorkChain._wrap_bare_dict_inputs`
method which is supposed to wrap bare dictionaries in the inputs if the
corresponding port accepts a `Dict` node. However, it was checking that
the `valid_type` was equal to `Dict` instead of checking that the tuple
contains the `Dict` type. As a results, bare dictionaries were no longer
automatically wrapped in `Dict` node after the change mentioned above.

Here the logic is updated to check whether the `Dict` type is in the
`valid_type` tuple.