`StorageBackend`: Add the `initialise` method
The current interface of the `StorageBackend` provides two method to
control the initialisation and lifetime of a storage:

 * `migrate`
 * `_clear`

The `migrate` method will initialise a completely new storage or migrate
an already initialised backend to the latest schema version, whereas the
`_clear` method will clear all data from the storage. In reality, the
current implementation of `_clear` will actually completely destroy the
container and recreate it from scratch, which causes a new repository
UUID to be generated.

Conceptually, but also practically, it is beneficial to restructure this
design slightly. When using storage for a testing profile, it may be
necessary to completely reset a storage backend and reinitialise it.

Therefore the `initialise` method is added to the `StorageBackend`
abstract base class. This method replaces the `migrate` method as the
method to be called when a new storage backend is created and needs to
be initialised. It should make sure that at the end of the initialisation
the storage is at the latest schema version, i.e., by calling `migrate`
internally.

The method takes an argument `reset`, which is `False` by default, but
when set to `True`, the storage backend should be completely reset before
reinitialising. Typically this entails that the storage is completely
destroyed and reinitialised in a fresh state.

This is used by the `TemporaryProfileManager` test utility class when
a new temporary profile is created. Although the tests in `aiida-core`
essentially always use the `PsqlDosBackend` and so the Postgres database
will always have the correct initialisation, storage backends are now
pluginable and in plugins we may want to switch between storage backends
within a test suite, in which case the storage backends needs to be fully
reinitialised.