`PsqlDosMigrator`: Remove hardcoding of table name in database reset
The `PsqlDosMigrator.reset_database` and `PsqlDosBackend._clear` methods
both deleted (almost) all database tables. The tables to delete were
hard-coded in a list by each method.

The `delete_tables` method is added to the `PsqlDosMigrator` class which
will delete all tables of the schema which are determined dynamically by
reflecting the metadata schema. The `exclude_tables` argument can be
used to define table names to exclude.

The new method is used by both classes in order to remove the
hard-coding of table names.