feat: reduce user interaction in `quicksetup`
When `verdi quicksetup` detected that a database user or database already
exists, it would prompt the user whether to use those and, if not, to pick a
new name.

Novice users likely do not know what exactly this choice entails;
furthermore, choosing to reuse the database user may actually not be
possible if the correct password for this user is not known in the
`config.json`.

Furthermore, if the user decides not to reuse the db user/db, there is
no point in letting them choose the new name. It is much easier to
autogenerate it (if the user really wants to choose the name, they can
pass it in via the `--db-username`/`--db-name` options to `verdi
quicksetup`).

Here, we remove any user interaction from `verdi quicksetup` and simply
generate new names for the dbuser and the dbname as needed.